+++ src/onegov.gazette/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds billing address to notices.
+  [msom]
+
 1.10.5 (2017-11-27)
 ~~~~~~~~~~~~~~~~~~~